# openapi/api-v1.yaml
openapi: 3.0.3
info:
  title: EsquiloSpeak API
  description: API specifications for the EsquiloSpeak MVP language learning app.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local API Gateway
security:
  - BearerAuth: []
paths:
...
  /courses/{courseId}/home:
    get:
      summary: Course dashboard/home data
      description: Get active unit, lessons, review status, and daily goals.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
      responses:
        '200':
          description: Course homepage details
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseId:
                    type: string
                    example: "en_for_vi"
                  activeUnitId:
                    type: string
                    example: "unit_001"
                  activeUnitTitle:
                    type: string
                    example: "Chào hỏi căn bản"
                  progressPercent:
                    type: integer
                    example: 35
                  lessons:
                    type: array
                    items:
                      type: object
                      properties:
                        lessonId:
                          type: string
                          example: "lesson_101"
                        title:
                          type: string
                          example: "Xin chào và Tạm biệt"
                        status:
                          type: string
                          enum: [locked, available, completed]
                          example: "available"
                  dueReviewCount:
                    type: integer
                    example: 5
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /courses/{courseId}/lessons/{lessonId}:
    get:
      summary: Get lesson details and questions
      description: Returns questions and assets within a lesson.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
          example: "lesson_101"
      responses:
        '200':
          description: Lesson details
          content:
            application/json:
              schema:
                type: object
                properties:
                  lessonId:
                    type: string
                    example: "lesson_101"
                  lessonVersionId:
                    type: string
                    example: "lesson_101_v1"
                  title:
                    type: string
                    example: "Xin chào và Tạm biệt"
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        questionId:
                          type: string
                          example: "q_201"
                        versionId:
                          type: string
                          example: "q_201_v1"
                        type:
                          type: string
                          enum: [multiple_choice, listening, translate]
                          example: "multiple_choice"
                        prompt:
                          type: string
                          example: "Dịch từ: 'Hello'"
                        audioUrl:
                          type: string
                          example: "http://localhost:8080/api/v1/media/hello.mp3"
                        correctAnswer:
                          type: string
                          example: "Xin chào"
                        explanation:
                          type: string
                          example: "'Hello' dịch sang tiếng Việt nghĩa là 'Xin chào'."
                        options:
                          type: array
                          items:
                            type: object
                            properties:
                              optionId:
                                type: integer
                                example: 1
                              optionText:
                                type: string
                                example: "Xin chào"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Standard Error Response
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "BAD_REQUEST"
                  message:
                    type: string
                    example: "Invalid input values"
                  details:
                    type: object
                    example: {}
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "UNAUTHORIZED"
                  message:
                    type: string
                    example: "Token is expired or invalid"
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Nhập JWT access token để gọi các API được bảo vệ.